+++ src/utils/data.ts
@@ -1,5 +1,10 @@
 import _ from 'lodash/fp'
-import { calculateTextFontSize, fitContentInsideContainer, translateShape } from './transform'
+import {
+  calculateTextFontSize,
+  fitContentInsideContainer,
+  roundForGrid,
+  translateShape
+} from './transform'
 import type {
   DrawablePicture,
   Point,
@@ -10,6 +15,7 @@ import type {
 import { fetchAndStringify, getBase64Image } from './file'
 import type { CustomTool } from 'types/tools'
 import { DEFAULT_SHAPE_PICTURE } from 'constants/tools'
+import type { GridFormatType } from 'constants/app'
 
 export const createPicture = (
   fileOrUrl: File | string,
@@ -69,8 +75,13 @@ export const createPicture = (
 export const createShape = (
   ctx: CanvasRenderingContext2D,
   shape: CustomTool,
-  cursorPosition: Point
+  cursorPosition: Point,
+  gridFormat: GridFormatType
 ): DrawableShape | undefined => {
+  const roundCursorPosition: Point = [
+    roundForGrid(cursorPosition[0], gridFormat),
+    roundForGrid(cursorPosition[1], gridFormat)
+  ]
   switch (shape.type) {
     case 'brush':
       shape.settings
@@ -78,7 +89,7 @@ export const createShape = (
         toolId: shape.id,
         type: shape.type,
         id: _.uniqueId(`${shape.type}_`),
-        points: [[cursorPosition]],
+        points: [[roundCursorPosition]],
         rotation: 0,
         style: {
           globalAlpha: shape.settings.opacity.default,
@@ -92,7 +103,7 @@ export const createShape = (
         toolId: shape.id,
         type: shape.type,
         id: _.uniqueId(`${shape.type}_`),
-        points: [cursorPosition, cursorPosition],
+        points: [roundCursorPosition, roundCursorPosition],
         rotation: 0,
         style: {
           globalAlpha: shape.settings.opacity.default,
@@ -109,7 +120,7 @@ export const createShape = (
         id: _.uniqueId(`${shape.type}_`),
         points: _.flow(
           _.range(0),
-          _.map(() => cursorPosition)
+          _.map(() => roundCursorPosition)
         )(shape.settings.pointsCount.default),
         rotation: 0,
         style: {
@@ -128,7 +139,7 @@ export const createShape = (
         id: _.uniqueId(`${shape.type}_`),
         points: _.flow(
           _.range(0),
-          _.map(() => cursorPosition)
+          _.map(() => roundCursorPosition)
         )(shape.settings.pointsCount.default),
         rotation: 0,
         style: {
@@ -146,8 +157,8 @@ export const createShape = (
         toolId: shape.id,
         type: shape.type,
         id: _.uniqueId(`${shape.type}_`),
-        x: cursorPosition[0],
-        y: cursorPosition[1],
+        x: roundCursorPosition[0],
+        y: roundCursorPosition[1],
         width: 1,
         height: 1,
         rotation: 0,
@@ -173,8 +184,8 @@ export const createShape = (
         toolId: shape.id,
         type: shape.type,
         id: _.uniqueId(`${shape.type}_`),
-        x: cursorPosition[0],
-        y: cursorPosition[1],
+        x: roundCursorPosition[0],
+        y: roundCursorPosition[1],
         value: defaultValue,
         fontSize,
         width: 50,
@@ -191,8 +202,8 @@ export const createShape = (
         toolId: shape.id,
         type: shape.type,
         id: _.uniqueId(`${shape.type}_`),
-        x: cursorPosition[0],
-        y: cursorPosition[1],
+        x: roundCursorPosition[0],
+        y: roundCursorPosition[1],
         radiusX: 0,
         radiusY: 0,
         rotation: 0,
@@ -209,8 +220,8 @@ export const createShape = (
         toolId: shape.id,
         type: shape.type,
         id: _.uniqueId(`${shape.type}_`),
-        x: cursorPosition[0],
-        y: cursorPosition[1],
+        x: roundCursorPosition[0],
+        y: roundCursorPosition[1],
         radius: 0,
         rotation: 0,
         style: {
@@ -224,9 +235,9 @@ export const createShape = (
   }
 }
 
-export const copyShape = (shape: DrawableShape) => {
+export const copyShape = (shape: DrawableShape, gridFormat: GridFormatType) => {
   return {
-    ...translateShape([20, 20], shape, [0, 0]),
+    ...translateShape([20, 20], shape, [0, 0], gridFormat),
     id: _.uniqueId(`${shape.type}_`)
   } as DrawableShape
 }